requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.0
web3==6.15.0
cryptography==41.0.7
requests==2.31.0
aiohttp==3.9.1
Pillow==10.0.1
